requirements.txt
Django==4.2.8

django-filter==23.5

djangorestframework==3.14.0